OCPBUGS-1443: Avoid using a lister before client caches sync
Previously, CVO could use a `ClusterVersion` lister before its
informer's cache was synced, leading to the following noise in a CVO
log, that could also propagate into an `Upgradeable=False` condition
and cause user-observable noise:

```
Unable to get ClusterVersion, err=clusterversion.config.openshift.io "version" not found
```

This is coming from one of the upgradeable checks called inside
`setUpgradeableCondition`. This method is called in two scenarios:
- in `ConfigMap` informer create/update/delete handlers
- in `syncUpgradeable->`upgradeableSync` path that handles work items
  from the `upgradeableQueue` queue. This path throttles the sync and
  only sets the conditions if sufficient time passed since the last
  sync

The failure case is that while caches of all informers are synced, the
`ConfigMap` informer handler can be called while the `ClusterVersion`
informer cache is not synced yet (so this is a race condition bug).

My original solution was to stop calling `setUpgradeableCondition` from
the handlers entirely, and queue the full upgradeable sync instead. This
fixes the issue at startup, but then it causes CVO to handle admin-ack
CM changes by going through the potentially throttled full sync path,
which (I think) could mean that CVO would react on manual admin action
with a non-trivial delay, which is poor UX (people expect their actions
to have observable effects). So instead, I made the handler checks
whether the informers are synced, and if they are, then sync the
conditions immediately, but if they are not, queue a full upgradeable
sync.

I'm not 100% convinced the increased complexity is worth it for a mostly
cosmetic fix, so I'm open to opinions.