tests: fix rare nil dereference in `TestCache`
Looking at `Cache.Match()` implementation, the `c.Condition.Match()` on L133 can be called with `targetCondition=nil` and that seems to be intentional. In tests, `c.Condition` is a mock, which did not expect that pointer to be `nil` and dereferenced it, resulting in an occassional panic in tests like https://github.com/openshift/cluster-version-operator/pull/871#issuecomment-1348927131. The fix stores the full pointer (its deepcopy to avoid sharing that memory) in the mock instead.

I needed to restructure the tests a little to allow adding a nil parameter testcase