pkg/cvo/upgradeable: refactor throttling
Refactor the code that handles throttling upgradeability checks. Create
a new method that computes the duration for which the existing
`Upgradeable` status is considered recent enough to not be synced, and
simply pass this duration to the `RecentlyChanged` method. The new
method is now unit tested, too. Upgradeable-related intervals are now
uncoupled to unrelated sync intervals and are grouped in a new struct.