lib/resourceapply/imagestream.go: Improve logging
This commit improves the logging of handling the ImageStream resources.
Previously, the logged `diff` wasn't helpful, and it didn't show the
reason why the update of the resource happened. Now, CVO logs the
different fields that caused the update. This change is similar
to the change introduced by the commit [1].

This commit also fixes the incorrectly written message when updating
an ImageStream. Previously, the CVO would log "Updating Namespace" even
though it would update an ImageStream.

[1] https://github.com/openshift/cluster-version-operator/commit/c58385eea897b5af790e4dd36103e978e7e4c597