lib/resourcemerge: Add defaulting for ImageStream
This commit will add CVO-side defaulting for the ImageStream resources.

The ImportMode field was introduced in 4.12, and its default value
is `Legacy` [1]. This change will fix CVO hotlooping on the ImageStream
resources because previously, the CVO wasn't aware of this default value.

The commit will also add defaulting for the referencePolicy field. Even
though the CVO doesn't reconcile this field in resources, it is added as
to keep the `ensureTagReferencev1Defaults` function up to date for
future use. Its default value is `Source` as mentioned in
the documentation [2].

The written tests are ensuring that the CVO properly only reconciles
the `lookupPolicy` field and required fields in reference tags. These
tags should have matching `importPolicy` and `from` fields because the
other fields are set by a different controller [3]. These tests are also
ensuring that any missing required reference tags are added.
The TestEnsureTagReferencev1Defaults was written to ensure that
the defaulting is working properly since CVO doesn't reconcile all of
the fields that have default values.

[1] https://docs.openshift.com/container-platform/4.12/release_notes/ocp-4-12-release-notes.html#ocp-4-12-images
[2] https://docs.openshift.com/container-platform/4.12/rest_api/image_apis/imagestream-image-openshift-io-v1.html#spec-tags-referencepolicy
[3] https://github.com/openshift/cluster-version-operator/commit/10b7b32b231cae15ad635fa66c2bdf3d486308d8