lib/resourcemerge: Fix removing CA bundle from webhooks in ValidatingWebhookConfiguration
This commit will fix CVO's behavior of removing the `caBundle` field of
webhooks inside ValidatingWebhookConfiguration resources when
reconciling manifests as described in the bug [1].

Currently, the users can inject the CA bundle into specific resources
via the service-ca operator by setting the
`service.beta.openshift.io/inject-cabundle=true` annotation in the
resources [2].

However, the CVO is not aware of this feature in all the respective
resources. CVO, when reconciling such a resource, can see an empty
`caBundle` field in the given release manifest, and CVO then removes any
value in the existing applied resource so it matches the release
manifest semantically.

This commit makes the CVO look into the existing resource and check for
the `service.beta.openshift.io/inject-cabundle=true` annotation. If it's
set then the service-ca controller is responsible for the CA bundle
fields, and CVO will ignore any differences between the existing
`caBundle` fields and the `caBundle` fields in the given release
manifest of the CRD.

This bug also caused the CVO to hotloop on
ValidatingWebhookConfiguration as mentioned in the bug [3].

[1] https://issues.redhat.com/browse/OCPBUGS-4015
[2] https://docs.openshift.com/container-platform/4.12/security/certificates/service-serving-certificate.html
[3] https://issues.redhat.com/browse/OCPBUGSM-44532