install: Explicitly mention 1 replica in ClusterVersionOperatorDown
To make it easier to self-serve recovery from "another admin had
scaled this to zero".  It doesn't really matter what >0 value is
chosen, because any created pods will figure out a leader lease, the
leasing leader will resume management of the Deployment, and adjust it
to 'replicas: 1'.  But that doesn't seem to be worth explaining, when
we can just suggest 1, and avoid distracting folks with ambiguity.