Bug 2090680: pkg/cvo/updatepayload.go: timeout payload retrieval
When we separated payload load from payload apply (#683) the context
used for the retrieval changed as well. It went from one that was
constrained by syncTimeout (2 -4 minutes) [1] to being the unconstrained
shutdownContext [2]. However if "force" is specified we explicitly set a
2 minute timeout in RetrievePayload. This commit creates a new context
with a reasonable timeout for RetrievePayload regardless of "force".

[1]
https://github.com/openshift/cluster-version-operator/blob/57ffa7c610fb92ef4ccd9e9c49e75915e86e9296/pkg/cvo/sync_worker.go#L605

[2]
https://github.com/openshift/cluster-version-operator/blob/57ffa7c610fb92ef4ccd9e9c49e75915e86e9296/pkg/cvo/cvo.go#L413