RetrievePayload: Improve timeouts for corner cases
The `RetrievePayload` performs two operations: verification and
download. Both can take a non-trivial amount of time to terminate, up to
"hanging" where CVO needs to abort the operation. The verification
result can be ignored when upgrade is forced. The CVO calls
`RetrievePayload` with a context that does not set a deadline, so
`RetrievePayload` previously set its own internal deadline, common for
both operations. This led to a suboptimal behavior on forced upgrades,
where "hanging" verification could eat the whole timeout budget, got
cancelled but its result was ignored (because of force). The code
tried to proceed with download but that immediately aborts because of
the expired context.

Improve timeouts in `RetrievePayload` for both input context states:
with and without deadline. If the input context sets a deadline, it is
respected. If it does not, the default, separate deadlines are applied
for both operations. In both cases, the code makes sure the hanging
verification never spends the whole budget. When verification terminates
fast, the rest of its alloted time is provided to the download
operation.