pkg/cvo: Set 'controller' in our ownerReferences
We've been setting ownerReferences on resources we create or manage
since 4dbb32741a (Added ownerReference to all objects created and
managed by CVO, 2021-06-29, #617).  But when we are actively
controlling a resource, we should be setting 'controller' to conform
with:

  $ grep -A100 'type OwnerReference struct' vendor/k8s.io/apimachinery/pkg/apis/meta/v1/types.go | grep -B3 Controller
          UID types.UID `json:"uid" protobuf:"bytes,4,opt,name=uid,casttype=k8s.io/apimachinery/pkg/types.UID"`
          // If true, this reference points to the managing controller.
          // +optional
          Controller *bool `json:"controller,omitempty" protobuf:"varint,6,opt,name=controller"`

With this commit, I'm moving the ClusterVersion owner to 'controller:
true' for resources we could possibly be comfortable replacing from
scratch.  Although things like "I've deleted the openshift-etcd
namespace" might cause some disruption today, possibly including
bricking the cluster, and are not recommended.  They could still
theoretically be hardened in the future to support clean replacement.

I considered using 'controller: true' for create-only resources too,
because:

* We'll still re-create create-only resources, and
* We don't include the ClusterVersion owner at all when there is a
  competing owner.

But create-only resources are create-only because we expect the admin
or other actor to manage them after creation, so the cluster-version
operator cannot commit to replacing all of that possible post-creation
configuration state.  Actors who try and use 'controller: true' to
judge "I may be able to delete this and let the controller replace it"
should not try to delete create-only resources unless they understand
the resource well enough to replace any desired configuration state on
their own.