pkg/cvo/status: Failing is not more serious than Degraded
The outgoing text goes way back to the local Failing type in
7f5b7f42dd (conditions: Use a consistent constant for the Failing
condition, 2019-05-19, #191).  But ClusterVersion doesn't include
Degraded, and ClusterOperator don't set Failing, so we don't need a
relative-seriousness ranking.  In practice, a Degraded=True
ClusterOperator is one of several issues that could lead to a
Failing=True ClusterVersion, and when that's the only issue going on,
they clearly have the same severity.  When an Available=False
ClusterOperator feeds a Failing=True ClusterVersion, that would be
worse than a Degraded=True Available=True ClusterOperator.  And there
may also be issues like the CVO failing to reconcile a peripheral
change like an alert rule where ClusterVersion is Failing=True despite
the issue being less severe than many Degraded=True ClusterOperator
situations.