lib/resourceapply/batch.go: Improve logging
This commit will improve the logging when handling the Job resources.
Previously the logged diff wasn't helpful, and it didn't show the reason
why the update of the resource happened. Now, CVO logs the different
fields that caused the update. This change is similar to the change
introduced by the commit for improving logging when updating
Deployments [1].

[1] https://github.com/openshift/cluster-version-operator/commit/c58385eea897b5af790e4dd36103e978e7e4c597