lib: Fix hotlooping on CronJob resources
This commit will fix the hotlooping caused by missing defaulting when
applying the CronJob resources.

Previously, the CronJob resources were reconciled as generic resources.
This caused the CVO to hotloop because it wasn't aware of the default
values for some of the fields in CronJob resources. CVO would then try
to update these fields resulting in hotlooping where the CVO
continuously tries to update the fields. This causes unnecessary API
calls and unnecessary CVO logging.

By handling the CronJob resources as CronJob resources and not as
generic resources, we can implement further defaulting of the resources.
This results in CVO now understanding that some of the missing fields in
the manifest mean that these fields have implicitly set their default
value that is known to the CVO.

The default values were set by following the Kubernetes documentation [1].

[1] https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.26/#cronjobspec-v1-batch