replace global clustercondition registry with instance based approach
Globals cause problems for unit tests (demonstrated in some of the unit
tests) and prevent general reuse.  The anti-pattern is avoided in
kuberentes code and should be OCP as well.  This updates to providing a
configured condition registry without init blocks and provides a
standard clustercondition registry for each consumption.