syncStatus: Reduce verbosity when syncing nothing of interest
Status is synchronized every 15 seconds, and each call is currently
logged with a lot of metadata, including full `status`. The metadata
include words `error` and `Failure` which usually have zero values, but
the words get noticed by filters and people eyeballing the logs.

Because most of the time this event is not interesting, we can only log
it if we happen to sync something interesting, like a failure, an only
log every event on higher verbosity.