Add flags for risk evaluation of conditional updates in HyperShift
This commit will introduce new flags and logic to the CVO regarding its
PromQL target for risk evaluation of conditional updates with the CVO
being run in a hosted control plane in mind.

For the CVO to successfully access a service that provides metrics
(in the case of the CVO in a standalone OpenShift cluster it's the
thanos-querier service), it needs three things. It needs the service's
address, a CA bundle to verify the certificate provided by the service
to allow secure communication using TLS between the actors [1], and
the authorization credentials of the CVO. Currently, the CVO hardcodes
the address, the path to the CA bundle, and the path to the credentials
file.

This is not ideal, as CVO is starting to be used in other repositories
such as HyperShift [2].

A path to the service CA bundle file is added to allow explicitly
setting the CA bundle file of the given query service.

Currently, the CVO is using a kube-api server to resolve the IP address
of a specific service in the cluster [3]. Add new flags that allow
configuring the CVO to resolve the IP address via DNS when DNS is
available to the CVO. This is the case for hosted CVOs in HyperShift.
The alternative in HyperShift would be to use the management cluster
kube-api server and give the hosted CVOs additional permissions.

Add flags to specify the PromQL target URL. This URL contains the used
scheme, port, and the server's name used for TLS configuration. In the
case, DNS is enabled, the URL is also used to query the Prometheus
server. In the case, the DNS is disabled, the server can be specified
via the Kubernetes service in the cluster that exposes the server.

A flag to specify the path to the credentials file was added for more
customizability. This flag also enables the CVO to not set the token
when it's not needed. A CVO can communicate with a Prometheus server
over HTTP. In the case, the token is not needed, it would be undesirable
for the CVO to send its token without reason over HTTP.

This commit also adds a new flag to specify whether the CVO resides in
a hosted control plane. In this case, the CVO will inject its cluster
Id into PromQL queries to differentiate between multiple time series
belonging to different hosted clusters [4].

	[1] https://docs.openshift.com/container-platform/4.12/security/certificate_types_descriptions/service-ca-certificates.html
	[2] https://github.com/openshift/hypershift
	[3] https://github.com/openshift/cluster-version-operator/pull/920
	[4] https://github.com/openshift/cincinnati-graph-data/pull/3591