lib: Refactor processing Job resources
The commit will move the logic of applying the spec of the Job
resources, the commit fixes a typo in a comment and simplifies the
condition in the `DeleteJobv1()` function so we can return early rather
than nest deeper.

This commit breaks up the function `EnsureJob` into `ensureJobSpec` and
`EnsureJob`. The new function can be then reused in the future.

For example, the `batchv1.JobSpec` structure is also present in the
structure `batchv1.CronJobSpec.JobTemplate`, and now the logic can be
reused here as well [1].

[1] https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.26/#jobtemplatespec-v1-batch