lib: Update applying Job resources
Although there is currently no hotlooping present when applying Job
resources, it is necessary to update it to fix hotlooping on CronJob
resources.

Some of the hotlooping on CronJob resources is caused by hotlooping on
the `CronJob.CronJobSpec.JobTemplate.Spec`field. To ensure it's
correctly implemented and to properly reuse existing code, this commit
updates applying of Job resources because its logic can be then reused
in handling CronJob resources. This also fixes the applying of Job
resources to prevent any hotlooping caused by missing defaulting and
missing applying for new fields in the future.

The default values were set by following the Kubernetes documentation [1].

[1] https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.26/#job-v1-batch