Dockerfile: bump to 1.19 and use public UBI pullspec
This Dockerfile is not used in CI [1] nor in product builds [2], so it may as
well be external contributor friendly and do not contain private CI
registry pullspecs.

We should use 1.20 but that image does not exist yet [3]. I do not see
any reason why to keep using UBI8 in this Dockerfile, so I bumped it to
UBI9.

[1] https://github.com/openshift/release/blob/7d59d083a7aa5f2aec28e807681a4b4e9ccdd088/ci-operator/config/openshift/cluster-version-operator/openshift-cluster-version-operator-master.yaml#L17-L25
[2] https://github.com/openshift-eng/ocp-build-data/blob/openshift-4.14/images/cluster-version-operator.yml#L3
[3] https://github.com/openshift/release/pull/40567