Use _ where event handlers don't use the object
This is to hint to any future developers that there is no existing
event object handling in the functions, and they need to add all of that
themselves, including (hopefully they realize) DeletedFinalStateUnknown
handling.  That's a fair bit of hope to get DeletedFinalStateUnknown in
place, but calling that specific aspect out in a comment or something
would risk them reading the comment and assuming we were giving them an
exhaustive list of guards to consider, when they may need additional
guards to match distant-future informer changes (they hopefully read
[1]).  And the amount of churn expected for our other delete handlers is
low, so investing even more time in trying to protect those future devs
seems uneccessary.

[1]:
https://pkg.go.dev/k8s.io/client-go/tools/cache#ResourceEventHandler

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>