availableupdates: do not reset lastTransitionTime on unchanged status (#964)
* availableupdates: do not reset lastTransitionTime on unchanged status

The code in `evaluateConditionalUpdates` correctly uses
`SetStatusCondition` to set conditions, which only updates the
`LastTransitionTime` field when `Status` differs between the original
and updated state. Previously though, the original state always
contained empty conditions, because conditional updates are always
obtained from OSUS and the fresh structure was never updated with
existing conditions from the in-cluster status.

* review: use existing mock condition instead of new code

* review: use real queue instead of a mock