pkg/cvo: Debug logging for the availableUpdatesQueue
I'm not entirely clear what the call stack is for this queue, so wrap
it with logging to make it easier to understand the queue lifecycle.

Stack construction is based on [1], but I'm concatenating the function
names to try to fit a bit of the stack on that one log line.

[1]: https://pkg.go.dev/runtime#example-Frames