resourcemerge: refactor EnsureSecurityContextConstraint
Clearly separate input/output. Do not use side effects on structures
passed as input, pass output only via return value. Cover both
`EnsureSecurityContextConstraint` and `ApplySecurityContextConstraintv1`
with tests, including the fact that `Apply...` does *not* call `Ensure...`
as expected, which is a bug reported as OCPBUGS-18386.

The new tests need additional files to be vendored, which is done in a
separate followup commit.