Properly reconcile SCC resources, tolerate cluster modifications
Previously CVO only created SCC resources when they did not exist, but
did not enforce any further state from the manifests afterwards, so e.g.
SCC changes on updates were never propagated to the cluster.

There is an existing `EnsureSecurityContextConstraint` method in the
code, but it was not used anywhere. It is possible that calling it from
`ApplySecurityContextConstraint` is all that is missing. The method has
a "merge" semantics on various lists though (so the resulting state is
the union of what is in the cluster and in the manifest), which plays
nice and tolerates user changes. This behavior will be removed in a
followup commit because we want CVO to strictly reconcile to the state
desired in the manifest.