resourcemerge: only reconcile `Volumes` in SCCs but track all changes
Because reconciling SCCs can endanger existing workloads, we only add
the merging semantics for the `Volumes` field, because we need to carry
updates we did in manifests. I have checked SCC manifests in 4.11, 4.12
and 4.13 payloads, and we only add items to `volumes`, so we do not need
other fields for now. For other fields, only track _whether_ we would
need to reconcile towards the manifest state, so that we can report this
cluster state to the admin before they upgrade to 4.14 which changes the
reconciliation behavior to strict.