availableupdates: populate internal state from CV status on startup
If the CVO is just starting up, it should populate its "known state" of
available updates from the `ClusterVersion` status, if its contains
them. Previous CVO may have evaluated the same graph data like the the
current one is about to do and if it did, there are likely existing
conditions in the status that we need to respect (for example, do not
bump a `lastTransitionTime` field on a condition on a conditinal update
that was already evaluated with the same result.