OCPBUGS-9050: Alert on failing conditional update risk evaluation (#961)
* OCPBUGS-9050: Alert on failing conditional update risk evaluation

* Addressed review feedback

- Used `max by` in the alert to drop irrelevant labels from alert
- Replaced further `"Recommended"` literals with new constant
- Do not `ToLower` booleanish string labels
- Report metrics labelled with `condition` and `status` instead of a
  specialized `recommended` label. We only export the `Recommended`
  condition status still.
- Naming and code structure tweaks

* Improve the help message for `cluster_version_conditional_update_condition_seconds`

* metrics: pivot `..._conditional_update_condition_seconds` from duration to ts

* availableupdates: populate internal state from CV status on startup

If the CVO is just starting up, it should populate its "known state" of
available updates from the `ClusterVersion` status, if its contains
them. Previous CVO may have evaluated the same graph data like the the
current one is about to do and if it did, there are likely existing
conditions in the status that we need to respect (for example, do not
bump a `lastTransitionTime` field on a condition on a conditinal update
that was already evaluated with the same result.