pkg/payload/precondition/clusterversion/rollback: Allow previous version within z-stream
From [1], Lala, Scott, and Subin all agreed that rollbacks to the
cluster's previous release should not be blocked, as long as the
previous and current version had the same major.minor version
(i.e. "within a z-stream", using x.y.z naming for major.minor.patch).
We agreed to ignore whether the previous history entry is Completed or
Partial, expecting that most users who are interested in rolling back
will make that decision at the first sign of trouble, and not after
attempting a few roll-forward retargets.

[1]: https://issues.redhat.com/browse/OTA-553