pkg/cvo/sync_worker: Verification-failure details for unforced updates too
We've including verification-failure details for forced updates since
5ad3c1434f (pkg/cvo/updatepayload: Event when forcing through a
sig-verification failure, 2022-04-07, 2022, #763), but had not been
including them in logs or other output in the "we aren't forcing, so
this blocks the update's acceptance" case.  This commit adds the
detail to the Event, so it's available, but keeps only the high-level
message in the RetrievePayload status output (which feeds the
ReleaseAccepted condition in ClusterVersion), because while the
low-level are useful for debugging, they're pretty chatty for
condition consumers that are more interested in just knowing basically
why the update request isn't being accepted.

The newline-to-// replacement is because apparently Event messages
truncate at the first newline.  I have not tracked down docs or source
to back that up, but confirmed it in pre-merge testing [1].

[1]: https://github.com/openshift/cluster-version-operator/pull/1003#issuecomment-1847671077