OCPBUGS-7714: yield lock while retrieving payload
Retrieving payload is a potentially long-running operation (download can
be slow or even timing out), and `RetrievePayload` is called while
holding the `SyncWorker` lock. Holding the lock for extensive amount
of time prohibits the manifest applying look from proceeding, resulting
in failures to reconcile state in the cluster.

`RetrievePayload` does not have access to anything protected by the
lock, so I believe we can safely yield the lock for its duration and
reacquire it afterwards.