sync_worker.go: Refactor `calculateNext` and pull up lock
Previously, `calculateNext` was a method on `SyncWorker`, locked its
lock, but only read its `work` member. It had a `SyncWork` parameter as
well which it modified.

Refactor the method so that the modified data is the receiver
and the read data is the parameter (this is less surprising than read
reciver and modified parameter). This allows us to pass only the read
member of `SyncWorker` inside, and lock this access at the callsite,
which already has one. `calculateNext` was not called from any other
place so this is safe.

This reduces the amount of methods that need to care about `SyncWorker`
locking, which is a little bit of a mess so simplifying is a good thing.