sync_worker.go: simplify `apply()` signature
`apply()` callsite creates a `statusWrapper` from a `w` content and
passes it to a method with `w` as a receiver, resulting in a weird
coupling of the receiver and the parameter. `apply()` can build the
`statusWrapper` itself, simplifying the signature and avoiding the
coupling. We just need to pass the previousStatus as a parameter (we
could avoid even that but we need to log previous status at the callsite
and we cannot call `Status()` twice b/c we could log something different
from what is actually used in `apply()`)