sync_worker.go: simplify `syncPayload` signature
Similarly to the previous commit, `loadUpdatedPayload` constructs a
`statusWrapper` reporter only to pass it into `w.syncPayload`, which is
its sole user. `syncPayload` has all data necessary to create the
instance itself, so there is no reason to make its signature more
complicated. I imagine this was done in the past for testing, in a
dependency injection fashion, but it seems not longer be the case.