status: extract status update logic to method
I would like to unit-test this logic but `syncStatus` does apiserver
operation so such tests would need a fake. Computing new status is
enough logic to constitute a method, so we can extract it so that we
can test it without a mock server.

Extracted with an IDE refactor helper.