status: uncouple `updateClusterVersionStatus` with `Operator`
`Operator` is a pretty beefy object with access to half of the universe,
but `updateClusterVersionStatus` only uses its member very little. It
reads its `release` member and calls its `getAvailableUpdates` method.

We can uncouple the long method by passing these two members as
parameters. It is slightly awkward for the `getAvailableUpdates` method
but I think the benefits outweight this.